requirements.txt
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.32.4
